+++ troncli/constants.py
@@ -2,6 +2,7 @@
 
 JAVA_TRON_RELEASES_URL = 'https://github.com/tronprotocol/java-tron/releases/download/'
 JAVA_TRON_LASTEST_VERSION = '3.2.1'
+JAVA_TRON_EVENT_NODE_ZIP_URL = 'https://github.com/tronprotocol/java-tron/archive'
 
 NODES_DIR = '/tron_nodes'